+++ src/App.js
@@ -1,6 +1,5 @@
 import "./App.css";
-import { Footer } from "./components/Footer/Footer";
-import { NavBar } from "./components/NavBar/NavBar";
+import { NavBar, Footer } from "./components";
 import { RoutersPath } from "./Routers/RoutersPath";
 
 function App() {
