+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export { AuthProvider } from "./auth-context";
+export { FilterProvider } from "./filter-context";
